Amended the document loader API to follow the relevant W3C standard.
The standard is described here: http://www.w3.org/TR/2003/WD-DOM-Level-3-LS-20030619/load-save.html#LS-DocumentLS

The main point is making .load() and .loadXML() part of the DOMDocument, and relying on event listeners over callbacks.

The actual meat of document loading will differ on client vs. server and so this has been broken out into a loader that is passed in with SceneDOM.setLoader().